[ORGA] Move assessment participant components
