+++ orga/app/components/participant/assessment/tabs.hbs
@@ -1,0 +1,18 @@
+<div class='panel campaign-details__controls'>
+  <nav class='navbar campaign-details-controls__navbar-tabs'>
+    <LinkTo
+      @route='authenticated.campaigns.assessment.results'
+      @models={{array @campaignId @participationId}}
+      class='navbar-item'
+    >
+      {{t 'pages.assessment-individual-results.tab.results'}}
+    </LinkTo>
+    <LinkTo
+      @route='authenticated.campaigns.assessment.analysis'
+      @models={{array @campaignId @participationId}}
+      class='navbar-item'
+    >
+      {{t 'pages.assessment-individual-results.tab.review'}}
+    </LinkTo>
+  </nav>
+</div>
