--- orga/app/components/routes/authenticated/campaign/assessment/details.hbs
@@ -1,93 +1,0 @@
-<article class="participant">
-  <header class="navigation">
-   <PreviousPageButton
-     @route="authenticated.campaigns.campaign.activity"
-     @routeId={{@campaign.id}}
-     @backButtonAriaLabel={{t 'common.actions.back'}}
-     aria-label={{t 'pages.campaign.name'}}
-   >
-    {{@campaign.name}}
-   </PreviousPageButton>
-  </header>
-
-  <section class="panel panel--header">
-    <header class="panel-header__headline">
-      <h2 class="panel-header-title">{{@campaignAssessmentParticipation.firstName}} {{@campaignAssessmentParticipation.lastName}}</h2>
-    </header>
-
-    <div class="panel-header__body">
-      <ul class="panel-header__data">
-        {{#if @campaignAssessmentParticipation.participantExternalId}}
-          <li class="panel-header-data__content">
-            <div class="label-text panel-header-data-content__label">{{@campaign.idPixLabel}}</div>
-            <div class="value-text panel-header-data-content__value">{{@campaignAssessmentParticipation.participantExternalId}}</div>
-          </li>
-        {{/if}}
-        <li class="panel-header-data__content">
-          <div class="label-text panel-header-data-content__label">{{t 'pages.campaign-individual-results.start-date'}}</div>
-          <div class="value-text panel-header-data-content__value">
-            {{moment-format @campaignAssessmentParticipation.createdAt 'DD MMM YYYY'}}
-          </div>
-        </li>
-        {{#unless @campaignAssessmentParticipation.isShared }}
-          <li class="panel-header-data__content">
-            <div class="label-text panel-header-data-content__label">{{t 'pages.assessment-individual-results.progression'}}</div>
-            <div class="value-text panel-header-data-content__value">
-              {{@campaignAssessmentParticipation.progression}}%
-            </div>
-          </li>
-        {{/unless}}
-        {{#if @campaignAssessmentParticipation.isShared }}
-          <li class="panel-header-data__content">
-            <div class="label-text panel-header-data-content__label">{{t 'pages.campaign-individual-results.shared-date'}}</div>
-              <div class="value-text panel-header-data-content__value">
-                {{moment-format @campaignAssessmentParticipation.sharedAt 'DD MMM YYYY'}}
-              </div>
-          </li>
-        {{/if}}
-      </ul>
-
-      {{#if @campaignAssessmentParticipation.isShared}}
-        <ul class="panel-header__data panel-header__data--highlight">
-          {{#if this.displayBadges}}
-            <li aria-label={{t 'pages.assessment-individual-results.badges'}} class="panel-header-data__content panel-header-data-content__badges">
-              <Badges @badges={{@campaignAssessmentParticipation.badges}}/>
-            </li>
-          {{/if}}
-          <li aria-label={{t 'pages.assessment-individual-results.result.label'}} class="panel-header-data__content panel-header-data-content__progress-bar">
-            <ProgressBar @value={{@campaignAssessmentParticipation.masteryPercentage}}>
-              {{t 'pages.assessment-individual-results.result.value' count=@campaignAssessmentParticipation.validatedSkillsCount total=@campaignAssessmentParticipation.targetedSkillsCount}}
-            </ProgressBar>
-          </li>
-          {{#if @campaign.hasStages}}
-            <li aria-label={{t 'pages.assessment-individual-results.stages.label'}} class="panel-header-data__content panel-header-data-content__stages">
-              <StageStars
-                @result={{@campaignAssessmentParticipation.masteryPercentage}}
-                @stages={{@campaign.stages}}
-                @withTooltip={{true}}
-                @tooltipPosition="bottom-left"
-              />
-            </li>
-          {{else}}
-            <li class="panel-header-data__content value-text value-text--highlight panel-header-data-content__mastery-percentage">
-              {{@campaignAssessmentParticipation.masteryPercentage}}%
-            </li>
-          {{/if}}
-        </ul>
-      {{/if}}
-    </div>
-  </section>
-
-  <div class="panel campaign-details__controls">
-    <nav class="navbar campaign-details-controls__navbar-tabs">
-      <LinkTo @route="authenticated.campaigns.assessment.results" class="navbar-item" @models={{array @campaign.id @campaignAssessmentParticipation.id}}>
-        {{t 'pages.assessment-individual-results.tab.results'}}
-      </LinkTo>
-      <LinkTo @route="authenticated.campaigns.assessment.analysis" class="navbar-item" @models={{array @campaign.id @campaignAssessmentParticipation.id}}>
-        {{t 'pages.assessment-individual-results.tab.review'}}
-      </LinkTo>
-    </nav>
-  </div>
-
-  {{outlet}}
-</article>
