--- orga/app/components/routes/authenticated/campaign/assessment/details.js
@@ -1,8 +1,0 @@
-import Component from '@glimmer/component';
-
-export default class Details extends Component {
-
-  get displayBadges() {
-    return this.args.campaign.hasBadges && this.args.campaignAssessmentParticipation.badges.length > 0;
-  }
-}
