+++ orga/app/templates/authenticated/campaigns/assessment.hbs
@@ -1,5 +1,15 @@
 {{page-title @model.campaign.name}}
-<Routes::Authenticated::Campaign::Assessment::Details
-  @campaign={{@model.campaign}}
-  @campaignAssessmentParticipation={{@model.campaignAssessmentParticipation}}
-/>
+
+<article class="participant">
+  <Participant::Assessment::Header
+    @campaign={{@model.campaign}}
+    @participation={{@model.campaignAssessmentParticipation}}
+  />
+
+  <Participant::Assessment::Tabs
+    @campaignId={{@model.campaign.id}}
+    @participationId={{@model.campaignAssessmentParticipation.id}}
+  />
+
+  {{outlet}}
+</article>
