+++ orga/app/templates/authenticated/campaigns/assessment/analysis.hbs
@@ -1,4 +1,5 @@
 {{page-title this.pageTitle}}
+
 <Campaign::Analysis::Recommendations
   @campaignTubeRecommendations={{@model.campaignAnalysis.campaignTubeRecommendations}}
   @displayAnalysis={{gt @model.campaignAnalysis.campaignTubeRecommendations.length 0}}
