+++ orga/app/templates/authenticated/campaigns/assessment/loading.hbs
@@ -1,2 +1,3 @@
 {{page-title (t "common.loading")}}
+
 <PixLoader />
