+++ orga/app/templates/authenticated/campaigns/assessment/results.hbs
@@ -1,8 +1,7 @@
 {{page-title this.pageTitle}}
-<Routes::Authenticated::Campaign::Assessment::Results
-  @campaignAssessmentParticipationFirstName={{@model.firstName}}
-  @campaignAssessmentParticipationLastName={{@model.lastName}}
-  @campaignAssessmentParticipationResult={{@model.campaignAssessmentParticipationResult}}
+
+<Participant::Assessment::Results
+  @results={{@model.campaignAssessmentParticipationResult}}
   @displayResults={{gt @model.campaignAssessmentParticipationResult.sortedCompetenceResults.length 0}}
   @class="participant-results"
 />
